# src/GestorIA.Api/openapi/v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GestorIA API",
    "description": "SPEC-009. Money is a string with two decimals, dates are ISO-8601, errors are RFC 9457 problem+json.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/transactions/{id}/classify": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Stores the user's class for a movement. A later call replaces an earlier one.",
        "operationId": "classifyTransaction",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionClassification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      }
    },
...
    "/api/v1/profiles/{id}": {
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "One stored profile.",
        "operationId": "getProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      },
      "put": {
        "tags": [
          "profiles"
        ],
        "summary": "Replaces a stored profile with the one in the body.",
        "operationId": "replaceProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileInputDocument"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileView"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "profiles"
        ],
        "summary": "Deletes the profile and everything stored for it. Nothing is kept: export it first to keep a copy.",
        "operationId": "deleteProfile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      }
    },
...
    "/api/v1/profiles/{id}/bank-statements": {
      "post": {
        "tags": [
          "transactions"
        ],
        "summary": "Imports a bank statement, the file itself as the body. A line an earlier import stored is not stored again.",
        "operationId": "importBankStatement",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "bank",
            "in": "query",
            "required": true,
            "schema": {
              "enum": [
                "bbva"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "The statement file as exported by the bank, at most 2097152 bytes.",
          "content": {
            "text/csv": {
              "schema": {
                "type": "string"
              }
            },
            "text/plain": {
              "schema": {
                "type": "string"
              }
            },
            "application/vnd.ms-excel": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BankStatementImport"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "413": {
            "description": "Payload Too Large",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "415": {
            "description": "Unsupported Media Type",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      }
    },
    "/api/v1/profiles/{id}/transactions": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "The stored movements in booking-date order: all of them, a year's, or a quarter's of that year.",
        "operationId": "listTransactions",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "year",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "quarter",
            "in": "query",
            "description": "Needs year.",
            "schema": {
              "$ref": "#/components/schemas/Quarter"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransactionView"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      }
    },
    "/api/v1/profiles/{id}/review-queue": {
      "get": {
        "tags": [
          "transactions"
        ],
        "summary": "The movements of the profile's tax year waiting for the user's class, unclear or with a rule's suggestion, in the transactions list's order.",
        "operationId": "getReviewQueue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReviewItem"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKey": [ ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "BankStatementImport": {
        "required": [
          "bank",
          "lines",
          "imported",
          "alreadyImported"
        ],
        "type": "object",
        "properties": {
          "bank": {
            "type": "string"
          },
          "lines": {
            "type": "integer",
            "format": "int32"
          },
          "imported": {
            "type": "integer",
            "format": "int32"
          },
          "alreadyImported": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "PreviousYearChoice": {
        "required": [
          "kind"
        ],
        "type": "object",
        "anyOf": [
          {
            "$ref": "#/components/schemas/PreviousYearChoiceNoActivityLastYear"
          },
          {
            "$ref": "#/components/schemas/PreviousYearChoicePreviousYearNet"
          }
        ],
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "noActivity": "#/components/schemas/PreviousYearChoiceNoActivityLastYear",
            "rendimientoNeto": "#/components/schemas/PreviousYearChoicePreviousYearNet"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "ProfileEmployment": {
        "required": [
          "ingresos",
          "seguridadSocial"
        ],
        "type": "object",
        "properties": {
          "ingresos": {
            "pattern": "^[0-9]{1,12}(\\.[0-9]{1,2})?$",
            "type": "string"
          },
          "seguridadSocial": {
            "pattern": "^[0-9]{1,12}(\\.[0-9]{1,2})?$",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
      "ProfileInputDocument": {
        "required": [
          "taxYear",
          "region",
          "employment",
          "activity",
          "projection"
        ],
        "type": "object",
        "properties": {
          "taxYear": {
            "type": "integer",
            "format": "int32"
          },
          "region": {
            "type": "string"
          },
          "employment": {
            "$ref": "#/components/schemas/ProfileEmployment"
          },
          "activity": {
            "$ref": "#/components/schemas/ProfileRegistration"
          },
          "projection": {
            "$ref": "#/components/schemas/ProfileProjection"
          }
        },
        "additionalProperties": false
      },
      "ProfileProjection": {
        "required": [
          "ingresos",
          "gastos",
          "baseCotizacion"
        ],
        "type": "object",
        "properties": {
          "ingresos": {
            "pattern": "^[0-9]{1,12}(\\.[0-9]{1,2})?$",
            "type": "string"
          },
          "gastos": {
            "pattern": "^[0-9]{1,12}(\\.[0-9]{1,2})?$",
            "type": "string"
          },
          "baseCotizacion": {
            "pattern": "^[0-9]{1,12}(\\.[0-9]{1,2})?$",
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ProfileRegistration": {
        "required": [
          "alta",
          "previousYear",
          "newActivity"
        ],
        "type": "object",
        "properties": {
          "alta": {
            "type": "string",
            "format": "date"
          },
          "previousYear": {
            "$ref": "#/components/schemas/PreviousYearChoice"
          },
          "newActivity": {
            "$ref": "#/components/schemas/NewActivityChoice"
          }
        },
        "additionalProperties": false
      },
      "ProfileView": {
        "required": [
          "id",
          "taxYear",
          "region",
          "employment",
          "activity",
          "projection"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "taxYear": {
            "type": "integer",
            "format": "int32"
          },
          "region": {
            "type": "string"
          },
          "employment": {
            "$ref": "#/components/schemas/ProfileEmployment"
          },
          "activity": {
            "$ref": "#/components/schemas/ProfileRegistration"
          },
          "projection": {
            "$ref": "#/components/schemas/ProfileProjection"
          }
        }
      },
...
      "Quarter": {
        "enum": [
          "Q1",
          "Q2",
          "Q3",
          "Q4"
        ]
      },
...
      "ReviewItem": {
        "required": [
          "id",
          "bookingDate",
          "valueDate",
          "description",
          "amount",
          "suggestion"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "bookingDate": {
            "type": "string",
            "format": "date"
          },
          "valueDate": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "amount": {
            "pattern": "^-?\\d+\\.\\d{2}$",
            "type": "string"
          },
          "suggestion": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ReviewSuggestion"
              }
            ]
          }
        }
      },
      "ReviewSuggestion": {
        "required": [
          "class",
          "ruleId"
        ],
        "type": "object",
        "properties": {
          "class": {
            "$ref": "#/components/schemas/TransactionClass"
          },
          "ruleId": {
            "type": "string"
          }
        }
      },
...
      "TransactionClass": {
        "enum": [
          "activityIncome",
          "deductibleExpense",
          "socialSecurity",
          "aeatPayment",
          "employmentIncome",
          "savingsIncome",
          "ownTransfer",
          "personal"
        ]
      },
      "TransactionClassification": {
        "required": [
          "class"
        ],
        "type": "object",
        "properties": {
          "class": {
            "$ref": "#/components/schemas/TransactionClass"
          }
        }
      },
      "TransactionView": {
        "required": [
          "id",
          "bookingDate",
          "valueDate",
          "description",
          "amount",
          "balance"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "bookingDate": {
            "type": "string",
            "format": "date"
          },
          "valueDate": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "amount": {
            "pattern": "^-?\\d+\\.\\d{2}$",
            "type": "string"
          },
          "balance": {
            "pattern": "^-?\\d+\\.\\d{2}$",
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "description": "The local installation's API key (SPEC-009 §3). The health checks do not need it.",
        "name": "X-Api-Key",
        "in": "header"
      }
    }
  },
...
}